Rename movie-results to MovieResults
Sorry, that was opinionated of me. I follow Vercel/NextJS's naming style of pages/directories being hyphenated and function /component names being camelcase
